% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distance_traveled}
\alias{distance_traveled}
\title{Distance traveled within segments}
\format{
A dataframe with 43 observations of 4 variables
\describe{
  \code{sub_basin } Sub-basin within Penobscot River catchment \cr \cr
  \code{stocking_location } Stocking location \cr \cr
  \code{site_code } Code for site/stocking location \cr \cr
  \code{km_traveled } Distance traveled from stocking location to ocean in km \cr \cr
}
}
\source{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
\usage{
distance_traveled
}
\description{
Distance traveled by Atlantic salmon smolts within 
collection units or segments in the Penobscot River, Maine based
on Stevens et al. (2019).
}
\keyword{datasets}
