% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downstream_medway}
\alias{downstream_medway}
\title{Downstream data for Medway}
\format{
A dataframe with 7,589 observations of 4 variables
\describe{
  \code{flow_cfs } Value of flow (discharge) \cr \cr
  \code{cumulative_prob } Cumulative flow probability \cr \cr
  \code{flow_prob } Probability of individual flow value \cr \cr
  \code{smolt_surv_prob } Smolt survival probability given flow \cr \cr
}
}
\source{
Amaral S, Fay C, Hecker G, Perkins N. 2012. Atlantic salmon survival 
estimates at mainstem hydroelectric projects on the Penobscot River. 
Phase 3 Final Report. Alden Research Laboratory, Inc., Holden, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.
}
\usage{
downstream_medway
}
\description{
Alden research estimates of downstream survival from the 
Atlantic salmon Dam Impact analysis model v67. Estimates are based on 
in-river flow.
}
\keyword{datasets}
