% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downstream_orono}
\alias{downstream_orono}
\title{Downstream data for Orono}
\format{
A dataframe with 4,457 observations of 4 variables
\describe{
  \code{flow_cfs } Value of flow (discharge) \cr \cr
  \code{cumulative_prob } Cumulative flow probability \cr \cr
  \code{flow_prob } Probability of individual flow value \cr \cr
  \code{smolt_surv_prob } Smolt survival probability given flow \cr \cr
}
}
\source{
Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.
}
\usage{
downstream_orono
}
\description{
Alden research estimates of downstream survival from the 
Atlantic salmon Dam Impact analysis model v67. Estimates are based on 
in-river flow. This dataset was updated in 2020 following changes to the
hydrosystem associated with the Penobscot River Restoration Project.
}
\keyword{datasets}
