% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_ranks}
\alias{flow_ranks}
\title{Cumulative distribution ranks for flows by year}
\format{
A dataframe with 9858 observations of 6 variables
\describe{
  \code{GageID } US Geological Survey gage ID \cr \cr
  \code{Location } Gage location \cr \cr
  \code{Day } Ordinal day of year for measurement \cr \cr
  \code{cum_prob } Cumulative flow probability \cr \cr
  \code{Year } Year of measurement
  \code{flow } Flow measurement
}
}
\source{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
\usage{
flow_ranks
}
\description{
Cumulative distribution ranks for flows by year at
USGS gage stations in the Penobscot River watershed, Maine, USA.
}
\keyword{datasets}
