% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_ratios}
\alias{flow_ratios}
\title{Flow ratios data set}
\format{
A dataframe with 21 observations of 3 variables
\describe{
  \code{dam } Name of dam \cr \cr
  \code{GageID } USGS Gage used to determine dam-specific flow
  \code{flow_ratio } Ratio of flow from gage \cr \cr
}
}
\source{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
\usage{
flow_ratios
}
\description{
GageID and associated flow ratios by dam in the Penobscot
River based on Stevens et al. (2019).
}
\keyword{datasets}
