% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_flows.R
\name{get_annual_flows}
\alias{get_annual_flows}
\title{Get annual flows}
\usage{
get_annual_flows(year = 2020)
}
\arguments{
\item{year}{Year of simulation}
}
\value{
A vector with discharge estimates drawn from a cdf
}
\description{
Simulate correlated annual flows at USGS gage stations 
in the Penobscot River watershed based on cumulative probability 
distribution used by Stevens et al. (2019).
}
\references{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
