% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dam_passage.R
\name{get_dam_passage}
\alias{get_dam_passage}
\title{Get dam-specific survival rates}
\usage{
get_dam_passage(year = 2020, flow)
}
\arguments{
\item{year}{Year of simulation}

\item{flow}{A numeric vector of length 21 containing flow values (cfs) 
corresponding to dams in rows in \code{\link{flow_ratios}} and
the `upstream` argument of \code{\link{run_one_year}}.}
}
\value{
A vector with dam passage rates
}
\description{
Get correlated dam-specific downstream survival rates for 
smolts at each dam from built-in `downstream_` datasets or default values
based on triangular distributions or fixed rates used in Stevens et al. 
(2019). Values based on Amaral et al. (2012) and sampled following 
Nieland et al. (2013, 2015) and Nieland and Sheehan (2020), sampled using 
`EnvStats::rtri()`, or fixed to 1 for other dams based on 
Stevens et al. (2019).
}
\references{
Amaral S, Fay C, Hecker G, Perkins N. 2012. Atlantic salmon survival 
estimates at mainstem hydroelectric projects on the Penobscot River. 
Phase 3 Final Report. Alden Research Laboratory, Inc., Holden, MA.

Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.

Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
