% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stocking_data.R
\name{get_stocking_data}
\alias{get_stocking_data}
\title{Get historical stocking data}
\usage{
get_stocking_data(year)
}
\arguments{
\item{year}{A numeric vector with a single year included in 
\code{\link{stocking_data}}.}
}
\value{
A vector with dam passage rates
}
\description{
Get number of smolts stocked at various stocking sites by year
in the Penobscot River, Maine, USA. Uses user-specified year to query the
`stocking_data` and merge with `distance_traveled` based on site_code.
}
\references{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
