% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latent_estuary}
\alias{latent_estuary}
\title{Latent estuary survival coefficients for dams}
\format{
A dataframe with 8 observations of 4 variables
\describe{
  \code{n_dams } Number of dams passed \cr \cr
  \code{n_dams_std } Standardized value for number of dams \cr \cr
  \code{s_logit } Logit-scale estuary survival based on number of dams passed \cr \cr
  \code{s_real } Real-scale estuary survival based on number of dams passed \cr \cr
}
}
\source{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.

Stich, DS, JF Kocik, GB Zydlewski, and JD Zydlewski. 2015. Linking behavior,
physiology, and survival of Atlantic Salmon smolts during estuary 
migration. Marine and Coastal Fisheries: Dynamics, Management, 
and Ecosystem Science 7:68-86.
}
\usage{
latent_estuary
}
\description{
Latent estuary survival coeffficients for number of 
dams passed by fish based on Stich et al. (2015) as implemented
by Stevens et al. (2019).
}
\keyword{datasets}
