% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Matt.R
\name{make_Matt}
\alias{make_Matt}
\title{Create Mattawamkeag River smolt population}
\usage{
make_Matt(stocking, prod, sat)
}
\arguments{
\item{stocking}{A dataframe matching output of 
\code{link{get_stocking_data}}.}

\item{prod}{A named vector matching the `prod` argument specified in 
\code{\link{run_one_year}}}

\item{sat}{A named vector matching the `sat` argument specified in 
\code{\link{run_one_year}}}
}
\description{
Function used to create a dataframe containing river segments
and sites for stocking and migration as well as dams from built-in data sets,
user-specified arguments in \code{\link{run_one_year}}, and output of
\code{\link{get_stocking_data}} for a given `year` in the Mattawamkeag
River from Stevens et al. (2019).
}
\references{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
