% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort_per_km}
\alias{mort_per_km}
\title{Empirical cumulative distribution function for natural mortality
per kilometer}
\format{
A dataframe with 34 observations of 3 variables
\describe{
  \code{prop_lost_per_km } Proportion of fish lost per kilometer \cr \cr
  \code{cumulative frequency } Cumulative probability at observation \cr \cr
  \code{prob } Probability of occurrence for a given observation \cr \cr
}
}
\source{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
\usage{
mort_per_km
}
\description{
Empirical cumulative distribution function for natural mortality
per kilometer derived by Stevens et al. (2019).
}
\keyword{datasets}
