% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penPass_shiny.R
\name{penPass_shiny}
\alias{penPass_shiny}
\title{Run the PenPass model interactively with shiny}
\usage{
penPass_shiny()
}
\description{
Run the PenPass model interactively in a browser using the shiny
package for R. Allows users to view or download summary data from multiple
runs of a single scenario.
}
\examples{
# To open the shiny user-interface, run the following in R console:
# penPass_shiny()

}
\seealso{
run_one_year
}
