% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PISC}
\alias{PISC}
\title{Piscataquis River sub-basin}
\format{
A dataframe with 54 observations of 6 variables
\describe{
  \code{huc_collection_segment_or_damname } Location within Penobscot River catchment corresponding to the same column in \code{\link{distance_traveled}} \cr \cr
  \code{habitat_unit_code } Habitat unit code(s) or HUCs \cr \cr
  \code{available_habitat_units_or_segment_length } Available habitat units or segment length of collection segment or dam reach \cr \cr
  \code{point_stock_pct_segment } Percent distance into segment length to stock \cr \cr
  \code{saturation } Proportion of wild smolt habitat occupied
  \code{n_smolts } Number of smolts (zero or NA by default)
}
}
\source{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
\usage{
PISC
}
\description{
Dataframe containing sub-basin characteristics for 
collection segments and dam names in the Piscataquis River
(Stevens et al. 2019).
}
\keyword{datasets}
