% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_downstream_migration.R
\name{run_downstream_migration}
\alias{run_downstream_migration}
\title{Run downstream migration module}
\usage{
run_downstream_migration(mat)
}
\arguments{
\item{mat}{A dataframe matching output of \code{link{make_WPN_hazard}},
\code{link{make_EPN_hazard}}, \code{link{make_Matt_hazard}}, 
\code{link{make_PISC_hazard}} or \code{link{make_PN_hazard}}.}
}
\description{
Function used to create natural survival and dam passage 
survival hazards for a given `year` in the East Branch Penobscot River from 
Stevens et al. (2019).
}
\references{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
