% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_downstream_migration_pn.R
\name{run_downstream_migration_pn}
\alias{run_downstream_migration_pn}
\title{Run downstream migration module}
\usage{
run_downstream_migration_pn(mat, year)
}
\arguments{
\item{mat}{A dataframe matching \code{link{make_PN_hazard}}}

\item{year}{Year of simulation}
}
\description{
Function used to run downstream migration for lower mainstem
Penobscot River based on output of \code{link{make_WPN_hazard}},
\code{link{make_EPN_hazard}}, \code{link{make_Matt_hazard}}, and
\code{link{make_PISC_hazard}} based on Stevens et al. (2019).
}
\references{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
