% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one_year.R
\name{run_one_year}
\alias{run_one_year}
\title{Run one year in penPass model}
\usage{
run_one_year(
  year = 2017,
  downstream = list(seeboomook = NA, ripogenus = NA, north_twin = NA, quakish = NA,
    dolby = NA, east_millinocket = NA, medway = NA, matagamon = 1, guilford = NA,
    moosehead = NA, browns_mills = NA, howland = NA, mattaseunk = NA, west_enfield = NA,
    milford = NA, great_works = 1, gilman_falls = NA, stillwater = NA, orono = NA, veazie
    = 1, bangor_waterworks = 1),
  km_surv = NULL,
  prod = c(WPN_prod = 3, EPN_prod = 3, Matt_prod = 3, PISC_prod = 3, PN_prod = 3),
  sat = c(WPN_sat = 0, EPN_sat = 0, Matt_sat = 0, PISC_sat = 0, PN_sat = 0)
)
}
\arguments{
\item{year}{Year used for simulation. Must be 1970 - 2020.}

\item{downstream}{Downstream dam passage survival rates for smolts}

\item{km_surv}{Natural downstream survival of smolts per kilometer}
}
\description{
Function used to simulate downstream migration in a single
year for available timeseries (1970 - 2020). A wrapper for the `sample()` function for resampling the empirical 
cumulative distribution function included in the \code{\link{mort_per_km}}
data set.
}
