% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocking_data}
\alias{stocking_data}
\title{Stocking data for Penobscot River watershed}
\format{
A dataframe with 388 observations of 3 variables
\describe{
  \code{Year } Year of stocking \cr \cr
  \code{site_code } Site code corresponding to `stocking_site` in `distance_traveled`
  \code{n_stocked } Number of smolts stocked
}
}
\source{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
\usage{
stocking_data
}
\description{
Number of Atlantic salmon smolts stocked in each of 22 sites
by year.
}
\keyword{datasets}
