% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downstream_brownsmills}
\alias{downstream_brownsmills}
\title{Downstream data for Browns Mills}
\format{
A dataframe with 857 observations of 4 variables
\describe{
  \code{flow_cfs } Value of flow (discharge) \cr \cr
  \code{cumulative_prob } Cumulative flow probability \cr \cr
  \code{flow_prob } Probability of individual flow value \cr \cr
  \code{smolt_surv_prob } Smolt survival probability given flow \cr \cr
}
}
\source{
Nieland et al. (2013, 2015)
}
\usage{
downstream_brownsmills
}
\description{
Alden research estimates of downstream survival from the 
Atlantic salmon Dam Impact analysis model v67. Estimates are based on 
in-river flow.
}
\keyword{datasets}
