% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_flows.R
\name{get_annual_flows}
\alias{get_annual_flows}
\title{Get annual flows}
\usage{
get_annual_flows(year)
}
\arguments{
\item{year}{Year of simulation}
}
\value{
A vector with discharge estimates drawn from a cdf for each
dam based on gage stations used by Stevens et al. (2019). Not all
dams have or need flow estimates within the PenPass model (`NA`).
}
\description{
Simulate correlated annual flows at USGS gage stations 
in the Penobscot River watershed based on cumulative probability 
distribution used by Stevens et al. (2019).
}
\examples{
# Get annual flows by dam for 2012
get_annual_flows(year = 2012)

}
\references{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
