% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_flows.R
\name{get_annual_flows}
\alias{get_annual_flows}
\title{Get dam-specific survival rates}
\usage{
get_annual_flows(year = 2020)
}
\arguments{
\item{new_or_old}{A character string indicating whether to use `"new"` 
(Amaral et al. 2012, Nieland and Sheehan 2020) or `"old"` (Amaral et al. 2012, 
Nieland et al. 2020) flow-correlated probabilities of p_stillwater as well 
as flow-correlated survival at `milford`, `orono`, and `stillwater` dams.}
}
\value{
A vector with dam passage rates
}
\description{
Get correlated dam-specific downstream survival rates for 
smolts at each dam from built-in `downstream_` datasets or default values
based on triangular distribution. Values based on Amaral et al. (2012) and 
sampled following Nieland et al. (2013, 2015) and Nieland and Sheehan (2020),
or sampled using `EnvStats::rtri()` for other dams based on Stevens et al. 
(2019).
}
\references{
Amaral S, Fay C, Hecker G, Perkins N. 2012. Atlantic salmon survival 
estimates at mainstem hydroelectric projects on the Penobscot River. 
Phase 3 Final Report. Alden Research Laboratory, Inc., Holden, MA.

Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
in the Penobscot River Watershed, with a Focus on Weldon Dam. US Department of 
Commerce, Northeast Fisheries Science Center Reference Document 19-16, Woods 
Hole, MA.

Nieland JL, Sheehan TF, Saunders R. 2015. Assessing demographic effects of dams
on diadromous fish: a case study for Atlantic salmon in the Penobscot River, 
Maine. ICES Journal of Marine Science 72:2423–2437. 

Nieland JL, Sheehan TF, Saunders R, Murphy JS, Trinko Lake TR, Stevens JR. 2013. 
Dam Impact Analysis model for Atlantic salmon in the Penobscot River, Maine. US 
Department of Commerce, Northeast Fisheries Science Center Reference Document 
13-09, Woods Hole, MA.

Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
