% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_km_mort.R
\name{sim_km_mort}
\alias{sim_km_mort}
\title{Simulate natural, per kilometer survival}
\usage{
sim_km_mort(prop_lost_per_km, n, prob)
}
\arguments{
\item{prop_lost_per_km}{Proportion of fish lost per kilometer}

\item{n}{Number of values to sample}

\item{prob}{Sampling probabilities for cdf based on Stevens et al. (2019).}
}
\description{
Function used to simulate mortality per kilometer
from the empirical cumulative distribution function used by
Stevens et al. (2019).
}
\details{
A wrapper for the `sample()` function for resampling the empirical 
cumulative distribution function included in the \code{\link{mort_per_km}}
data set.
}
\examples{
# Simulate 
morts <- sim_km_mort(
  prop_lost_per_km = penPass::mort_per_km$prop_lost_per_km,
  n = 1000,
  prob = penPass::mort_per_km$prob
)

# Calculate number surviving from 1000 starting individuals
surv <- 1e3 - (1e3 * exp(-morts)) 
hist( surv, breaks = 40 )

}
\references{
Stevens, JR, JF Kocik, and TF Sheehan. 2019. Modeling the impacts of dams and 
stocking practices on an endangered Atlantic salmon (Salmo salar) 
population in the Penobscot River, Maine, USA. Canadian Journal of Fisheries
and Aquatic Sciences 76:1795-1807.
}
