% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stillwater_pars}
\alias{stillwater_pars}
\title{Stillwater logistic regression parameters}
\format{
A dataframe with 2 observations of 5 variables
\describe{
  \code{parameter } Model parameter \cr \cr
  \code{estimate } Estimated mean for parameter
  \code{se } Estimated standard error for parameter \cr \cr
  \code{z test statistic for regression}
  \code{p p-values for parameters}
}
}
\source{
Stich, DS, M Bailey, and JD Zydlewski. 2014. Survival of 
Atlantic Salmon Salmo salar smolts through a hydropower 
complex. Journal of Fish Biology 85:1074-1096.
}
\usage{
stillwater_pars
}
\description{
Parameter estimates and statistics for logistic regression
characterizing probability of fish using Stillwater Branch for downstream
migration around Marsh Island in the Penobscot River, USA. Used to predict
probability of fish using Stillwater Branch conditional on flow (cfs).
}
\keyword{datasets}
